Stub out SyncContainer Test and rename all other test functions
